JG-40 Feat: formatTime 함수 유틸화
